Implement prefix functions
Pool management may need a little bit more work. Specifically, I'm not sure if the idea of calling the EventListenerModule with the Discord.Client and Pool as arguments to create a new function that attaches to the event will work; I think it will, because it should pass Pool by reference, but I don't know.